Add files via upload
here is the html file for Django but I had to put this in the user files for some reason. C:\Users\DruSk\anaconda3\Lib\site-packages\django\contrib\admin\templates\weatherapp and in the C:\Users\DruSk\anaconda3\Lib\site-packages\django\contrib\auth\templates\weatherapp this is where I had to put it to make it work I think in the admin track I had to create the weatherapp folder to put this html file in.